requirements.txt
amqp==5.0.9
api-display-purposes==0.0.3
APScheduler==3.6.3
asgiref==3.5.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
clarifai==2.6.2
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
configparser==3.8.1
cryptography==36.0.1
decorator==4.4.2
Deprecated==1.2.13
Django==3.2.8
django-debug-toolbar==3.2.4
django-polymorphic==3.1.0
dnspython==2.2.0
EasyProcess==1.1
emoji==1.6.3
eventlet==0.33.0
flower==1.0.0
future==0.18.2
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
h11==0.13.0
huepy==1.2.1
humanize==4.0.0
idna==2.10
imageio==2.16.0
imageio-ffmpeg==0.4.5
iniconfig==1.1.1
instabot==0.117.0
instapy==0.6.16
jsonschema==2.6.0
kombu==5.2.3
MeaningCloud-python==2.0.0
mock==4.0.3
moviepy==1.0.3
numpy==1.22.2
outcome==1.1.0
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
plyer==2.0.0
proglog==0.1.9
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.0.1
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.1.4
regex==2022.1.18
requests==2.27.1
requests-toolbelt==0.9.1
responses==0.18.0
schedule==1.1.0
selenium==4.1.0
semantic-version==2.9.0
setuptools-rust==1.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webdriverdownloader==1.1.0.3
wrapt==1.13.3
wsproto==1.0.0
YooMoney==0.1.0
